Python3: remove long() and float(), use int()
Python 3 only has int() for integers, so use it in place of long().  Also,
numbers are float() by default so only make them int() when it is actually
needed.

+++ splat/scales.py
@@ -159,8 +159,8 @@ class HarmonicScale(Scale):
         # f = math.pow(k, n)
         # f = math.pow(self.base, math.log(f, self.base) % 1)
         coefs = list()
-        m = self._steps / 2
-        n = self._steps - m
+        m = int(self._steps / 2)
+        n = int(self._steps - m)
         for i in range(m):
             r = math.pow(k, i)
             while r > self.base:
